# backend/api/openapi.yaml
openapi: 3.0.0
info:
  description: Simple energyx API
  version: 3.0.0
  title: Simple energyx API
  contact:
    email: azdevteam@microsoft.com

components:
  schemas:
    energyxItem:
      type: object
      required:
        - listId
        - name
        - description
      description: A task that needs to be completed
      properties:
        id:
          type: string
        listId:
          type: string
        name:
          type: string
        description:
          type: string
        state:
          $ref: "#/components/schemas/energyxState"
        dueDate:
          type: string
          format: date-time
        completedDate:
          type: string
          format: date-time
    energyxList:
      type: object
      required:
        - name
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
      description: " A list of related energyx items"
    energyxState:
      type: string
      enum:
        - energyx
        - inprogress
        - done
  parameters:
    listId:
      in: path
      required: true
      name: listId
      description: The energyx list unique identifier
      schema:
        type: string
    itemId:
      in: path
      required: true
      name: itemId
      description: The energyx item unique identifier
      schema:
        type: string
    state:
      in: path
      required: true
      name: state
      description: The energyx item state
      schema:
        $ref: "#/components/schemas/energyxState"
    top:
      in: query
      required: false
      name: top
      description: The max number of items to returns in a result
      schema:
        type: number
        default: 20
    skip:
      in: query
      required: false
      name: skip
      description: The number of items to skip within the results
      schema:
        type: number
        default: 0

  requestBodies:
    energyxList:
      description: The energyx List
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/energyxList"
    energyxItem:
      description: The energyx Item
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/energyxItem"

  responses:
    energyxList:
      description: A energyx list result
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/energyxList"
    energyxListArray:
      description: An array of energyx lists
      content:
        application/json:
          schema:
            type: array
            items:
              $ref: "#/components/schemas/energyxList"
    energyxItem:
      description: A energyx item result
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/energyxItem"
    energyxItemArray:
      description: An array of energyx items
      content:
        application/json:
          schema:
            type: array
            items:
              $ref: "#/components/schemas/energyxItem"

paths:
  /lists:
    get:
      operationId: GetLists
      summary: Gets an array of energyx lists
      tags:
        - Lists
      parameters:
        - $ref: "#/components/parameters/top"
        - $ref: "#/components/parameters/skip"
      responses:
        200:
          $ref: "#/components/responses/energyxListArray"
    post:
      operationId: CreateList
      summary: Creates a new energyx list
      tags:
        - Lists
      requestBody:
        $ref: "#/components/requestBodies/energyxList"
      responses:
        201:
          $ref: "#/components/responses/energyxList"
        400:
          description: Invalid request schema
  /lists/{listId}:
    get:
      operationId: GetListById
      summary: Gets a energyx list by unique identifier
      tags:
        - Lists
      parameters:
        - $ref: "#/components/parameters/listId"
      responses:
        200:
          $ref: "#/components/responses/energyxList"
        404:
          description: energyx list not found
    put:
      operationId: UpdateListById
      summary: Updates a energyx list by unique identifier
      tags:
        - Lists
      requestBody:
        $ref: "#/components/requestBodies/energyxList"
      parameters:
        - $ref: "#/components/parameters/listId"
      responses:
        200:
          $ref: "#/components/responses/energyxList"
        404:
          description: energyx list not found
        400:
          description: energyx list is invalid
    delete:
      operationId: DeleteListById
      summary: Deletes a energyx list by unique identifier
      tags:
        - Lists
      parameters:
        - $ref: "#/components/parameters/listId"
      responses:
        204:
          description: energyx list deleted successfully
        404:
          description: energyx list not found
  /lists/{listId}/items:
    post:
      operationId: CreateItem
      summary: Creates a new energyx item within a list
      tags:
        - Items
      requestBody:
        $ref: "#/components/requestBodies/energyxItem"
      parameters:
        - $ref: "#/components/parameters/listId"
      responses:
        201:
          $ref: "#/components/responses/energyxItem"
        404:
          description: energyx list not found
    get:
      operationId: GetItemsByListId
      summary: Gets energyx items within the specified list
      tags:
        - Items
      parameters:
        - $ref: "#/components/parameters/listId"
        - $ref: "#/components/parameters/top"
        - $ref: "#/components/parameters/skip"
      responses:
        200:
          $ref: "#/components/responses/energyxItemArray"
        404:
          description: energyx list not found
  /lists/{listId}/items/{itemId}:
    get:
      operationId: GetItemById
      summary: Gets a energyx item by unique identifier
      tags:
        - Items
      parameters:
        - $ref: "#/components/parameters/listId"
        - $ref: "#/components/parameters/itemId"
      responses:
        200:
          $ref: "#/components/responses/energyxItem"
        404:
          description: energyx list or item not found
    put:
      operationId: UpdateItemById
      summary: Updates a energyx item by unique identifier
      tags:
        - Items
      requestBody:
        $ref: "#/components/requestBodies/energyxItem"
      parameters:
        - $ref: "#/components/parameters/listId"
        - $ref: "#/components/parameters/itemId"
      responses:
        200:
          $ref: "#/components/responses/energyxItem"
        400:
          description: energyx item is invalid
        404:
          description: energyx list or item not found
    delete:
      operationId: DeleteItemById
      summary: Deletes a energyx item by unique identifier
      tags:
        - Items
      parameters:
        - $ref: "#/components/parameters/listId"
        - $ref: "#/components/parameters/itemId"
      responses:
        204:
          description: energyx item deleted successfully
        404:
          description: energyx list or item not found
  /lists/{listId}/items/state/{state}:
    get:
      operationId: GetItemsByListIdAndState
      summary: Gets a list of energyx items of a specific state
      tags:
        - Items
      parameters:
        - $ref: "#/components/parameters/listId"
        - $ref: "#/components/parameters/state"
        - $ref: "#/components/parameters/top"
        - $ref: "#/components/parameters/skip"
      responses:
        200:
          $ref: "#/components/responses/energyxItemArray"
        404:
          description: energyx list or item not found
    put:
      operationId: UpdateItemsStateByListId
      summary: Changes the state of the specified list items
      tags:
        - Items
      requestBody:
        description: unique identifiers of the energyx items to update
        content:
          application/json:
            schema:
              type: array
              items:
                description: The energyx item unique identifier
                type: string
      parameters:
        - $ref: "#/components/parameters/listId"
        - $ref: "#/components/parameters/state"
      responses:
        204:
          description: energyx items updated
        400:
          description: Update request is invalid